# Presentation/Angular/src/app/api/models/product-available-dto-list-result.ts
/* tslint:disable */
/* eslint-disable */
import { ProductAvailableDto } from './product-available-dto';
export interface ProductAvailableDtoListResult {
  data?: null | Array<ProductAvailableDto>;
  message?: null | string;
}