# Presentation/Angular/src/app/api/models/product-available-dto.ts
/* tslint:disable */
/* eslint-disable */
export interface ProductAvailableDto {
  barcode?: null | string;
  description?: null | string;
  price?: number;
  productId?: null | string;
  stock?: number;
  type?: null | string;
}